Separate public interface for decompression from private implementation.
